requirements.txt
# External packages
appdirs
# pathlib2 is needed for Python 2.7 compatibility
pathlib2; python_version < "3"
pyyaml

# Microchip packages
pyedbglib>=2.21
pymcuprog>=3.13.7
